+++ packages/ui/src/components/index.ts
@@ -1,4 +1,4 @@
 export * from './Episodes/EpisodeCard';
 export * from './Image/NextImage';
-export * from './Layout/Fotter';
+export * from './Layout/Footer';
 export * from './Layout/Layout';
